+++ packages/responses/src/Domain/Item/CheckIntegrityResponse.ts
@@ -1,0 +1,8 @@
+import { IntegrityPayload } from '@standardnotes/payloads'
+import { MinimalHttpResponse } from '../Http/MinimalHttpResponses'
+
+export type CheckIntegrityResponse = MinimalHttpResponse & {
+  data: {
+    mismatches: IntegrityPayload[]
+  }
+}