+++ packages/services/src/Domain/Integrity/IntegrityEventPayload.ts
@@ -1,0 +1,7 @@
+import { RawPayload } from '@standardnotes/payloads'
+import { SyncSource } from '../Sync/SyncSource'
+
+export type IntegrityEventPayload = {
+  rawPayloads: RawPayload[]
+  source: SyncSource
+}