+++ packages/snjs/lib/services/Items/ItemManager.ts
@@ -14,6 +14,8 @@ import {
   PayloadSource,
   ItemInterface,
   predicateFromDSLString,
+  ItemManagerInterface,
+  IntegrityPayload,
 } from '@standardnotes/payloads'
 import { ItemCollectionNotesView } from '@Lib/protocol/collection/item_collection_notes_view'
 import { NotesDisplayCriteria } from '@Lib/protocol/collection/notes_display_criteria'
@@ -85,7 +87,11 @@ export const isTagOrNote = (x: SNItem): x is SNNote | SNTag =>
  * will then notify  its observers (which is us), we'll convert the payloads to items,
  * and then  we'll propagate them to our listeners.
  */
-export class ItemManager extends AbstractService implements ItemsClientInterface {
+export class ItemManager
+  extends AbstractService
+  implements
+    ItemManagerInterface,
+    ItemsClientInterface {
   private unsubChangeObserver: () => void
   private observers: Observer[] = []
   private collection!: ItemCollection
@@ -241,6 +247,10 @@ export class ItemManager extends AbstractService implements ItemsClientInterface
    */
   public get invalidItems() {
     return this.collection.invalidElements()
+  }
+
+  public get integrityPayloads(): IntegrityPayload[] {
+    return this.collection.integrityPayloads()
   }
 
   /**
@@ -1213,6 +1223,15 @@ export class ItemManager extends AbstractService implements ItemsClientInterface
   }
 
   public getFilesForNote(note: SNNote): SNFile[] {
-    return this.itemsReferencingItem(note.uuid).filter(ref => ref.content_type === ContentType.File) as SNFile[];
-  };
+    return this.itemsReferencingItem(note.uuid).filter(
+      (ref) => ref.content_type === ContentType.File,
+    ) as SNFile[]
+  }
+
+  public renameFile(file: SNFile, name: string, ext?: string): Promise<SNFile> {
+    return this.changeItem<FileMutator, SNFile>(file.uuid, (mutator) => {
+      mutator.name = name
+      mutator.ext = ext
+    })
+  }
 }